web: utils.saveModel() with common code for parsing and handling Upstream errors
