+++ web/js/packs.js
@@ -3,6 +3,7 @@
 var backbone = require('backbone');
 var React = require('react');
 var CollectionLoader = require('./components/CollectionLoader.jsx');
+var utils = require('./utils');
 require('react.backbone');
 
 var packs = (function () {
@@ -17,19 +18,13 @@ var Pack = React.createBackboneClass({
   changeOptions: 'change:amount',
 
   onSave: function () {
-    var changes = {
-      amount: parseInt(this.refs.amountInput.value, 10) || 0
-    };
-
-    this.getModel().save(changes, {
-      success: function () {
-        alert('succeeded');
-      },
-
-      error: function () {
-        alert('failed');
-      }
-    })
+    utils.saveModel(
+      this.getModel(),
+      {amount: parseInt(this.refs.amountInput.value, 10) || 0},
+      {},
+      { success: 'Pack Saved',
+        error: 'Failed to update Pack' }
+    );
   },
 
   render: function () {
